Get closer to working testing with old versions of Caffe2.
Since show_progress is a newly added attribute, it will fail
when run with an older version of Caffe2.  Catch this situation
and retry without the argument.

Also, it's wrong to exit with exit code zero when there's an error
condition.  Use a non-zero exit code.

This isn't enough because some downloaded model format shave changed in
BC-incompatible ways.  That is not going to be fixable.

Signed-off-by: Edward Z. Yang <ezyang@fb.com>