Rewrite translation of Conv/ConvTranspose/Pool ops
The new structure more closely mirrors how Caffe2 handles these
matters, which is good because it means that fixes in one
case will get applied in all cases.

This also adds a fix to handle 'pads' which are the same size as
'kernels': Caffe2 requires pads to be twice the size, so we
just multiply to match it up.

Signed-off-by: Edward Z. Yang <ezyang@fb.com>