ConvPoolBaseOp refactor (#22)
* Get closer to working testing with old versions of Caffe2.

Since show_progress is a newly added attribute, it will fail
when run with an older version of Caffe2.  Catch this situation
and retry without the argument.

Also, it's wrong to exit with exit code zero when there's an error
condition.  Use a non-zero exit code.

This isn't enough because some downloaded model format shave changed in
BC-incompatible ways.  That is not going to be fixable.

Signed-off-by: Edward Z. Yang <ezyang@fb.com>

* Rewrite translation of Conv/ConvTranspose/Pool ops

The new structure more closely mirrors how Caffe2 handles these
matters, which is good because it means that fixes in one
case will get applied in all cases.

This also adds a fix to handle 'pads' which are the same size as
'kernels': Caffe2 requires pads to be twice the size, so we
just multiply to match it up.

Signed-off-by: Edward Z. Yang <ezyang@fb.com>

* Undo unicode_literals for protobuf access.

Signed-off-by: Edward Z. Yang <ezyang@fb.com>

* Rename attr to arg.

Signed-off-by: Edward Z. Yang <ezyang@fb.com>

* AveragePool is a ConvPoolOp.

Signed-off-by: Edward Z. Yang <ezyang@fb.com>