backend.py: introduce ModelHelper [draft/WIP]
There's code in caffe2 that adds an LSTM to the net, but it requires
ModelHelper, which doesn't fit with the "return list of ops" approach
of backend.py. Therefore, a preliminary step to incorporating that
code is to thread ModelHelper through everywhere. Rather than fully
convert to ModelHelper, the various _create_* methods are now free to
either return a list of ops or to directly operate on model_helper.
_create_concat has been converted as an example.