Introduce Elided Arguments, and ignore input sequence_lens
while ONNX has a way to represent missing arguments (pass in empty
string), that's not representable in the graph that optimize-onnx and
pytorch use. Therefore, introduce a dummy node type that stands in for
such missing arguments.

Also, ignore the passed-in sequence lens argument in favor of
rederiving it from the input size. We still much provide something to
caffe2, since it is a required argument to lstm op.